# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "aptpay App API",
    "description": "Find out how your APIs work",
    "license": {
      "name": "tejas"
    }
  },
  "host": "uataptpay.moreyeahs.in",
  "basePath": "/api",
...
  "schemes": ["https", "http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "APIKeyHeader": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "security": [
    {
      "APIKeyHeader": []
    }
  ],
  "paths": {
...
    "/wallet/createWallet": {
      "post": {
        "summary": "walletCreate",
        "tags": ["wallet"],
        "description": "walletCreate",
        "parameters": [
          {
            "name": "walletCreate",
            "in": "body",
            "description": "walletCreate for user",
            "schema": {
              "$ref": "#/definitions/walletCreate"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/walletCreateResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/walletCreateErrorResponse"
            }
          }
        }
      }
    },
    "/wallet/addWalletBalance": {
      "post": {
        "summary": "add fund to wallet",
        "tags": ["wallet"],
        "description": "add fund to wallet",
        "parameters": [
          {
            "name": "add fund to wallet",
            "in": "body",
            "description": "add fund to wallet for user",
            "schema": {
              "$ref": "#/definitions/walletAdd"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/walletAddResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/walletAddErrorResponse"
            }
          }
        }
      }
    },
    "/wallet/sendWalletBalance": {
      "post": {
        "summary": "withdraw fund from wallet",
        "tags": ["wallet"],
        "description": "withdraw fund from wallet",
        "parameters": [
          {
            "name": "withdraw fund from wallet",
            "in": "body",
            "description": "withdraw fund from wallet",
            "schema": {
              "$ref": "#/definitions/walletWithdraw"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/walletWithdrawResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/walletWithdrawErrorResponse"
            }
          }
        }
      }
    },
    "/wallet/getWalletBalance": {
      "get": {
        "summary": "getWalletBalance",
        "tags": ["wallet"],
        "description": "getWalletBalance",

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/getWalletBalanceResponse"
            }
          },
          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/getWalletBalanceErrorResponse"
            }
          }
        }
      }
    },
    "/wallet/GetWalletTransactionHistory": {
      "get": {
        "summary": "Get Wallet Transaction History",
        "tags": ["wallet"],
        "description": "Get Wallet Transaction History",

        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/GetWalletTransactionHistoryErrorResponse"
            }
          },

          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/GetWalletTransactionHistory"
            }
          }
        }
      }
    },
    "/wallet/disableWallet": {
      "put": {
        "summary": "Disable Wallet",
        "tags": ["wallet"],
        "description": "Disable Wallet",
        "parameters": [
          {
            "name": "Disable Wallet",
            "in": "body",
            "description": "Disable Wallet",
            "schema": {
              "$ref": "#/definitions/disableWallet"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/disableWalletResponse"
            }
          },

          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/disableWalletErrorResponse"
            }
          }
        }
      }
    },
    "/wallet/enableWallet": {
      "put": {
        "summary": "Enable Wallet",
        "tags": ["wallet"],
        "description": "Enaable Wallet",
        "parameters": [
          {
            "name": "Enable Wallet",
            "in": "body",
            "description": "Enable Wallet",
            "schema": {
              "$ref": "#/definitions/enableWallet"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "  status: true ",
            "schema": {
              "$ref": "#/definitions/enableWalletResponse"
            }
          },

          "200": {
            "description": " status: false ",
            "schema": {
              "$ref": "#/definitions/enableWalletErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "walletCreate": {
      "properties": {
        "payeeId": {
          "type": "number"
        }
      }
    },
    "walletCreateResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "format": "string"
        },
        "data": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "walletCreateErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": false
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
    "getWalletBalanceResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "format": "string"
        },
        "data": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "getWalletBalanceErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": false
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
    "GetWalletTransactionHistory": {
      "properties": {
        "wallet": {
          "type": "string"
        }
      }
    },
...
    "GetWalletTransactionHistoryErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": false
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
    "walletAdd": {
      "properties": {
        "payeeId": {
          "type": "number"
        },
        "amount": {
          "type": "number"
        }
      }
    },
    "walletAddResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "format": "string"
        },
        "data": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "walletAddErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": false
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
    "walletWithdraw": {
      "properties": {
        "wallet": {
          "type": "string"
        },
        "payeeId": {
          "type": "number"
        },
        "amount": {
          "type": "number"
        },
        "referenceId": {
          "type": "string"
        }
      }
    },
    "walletWithdrawResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "format": "string"
        },
        "data": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "walletWithdrawErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": false
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
    "disableWallet": {
      "properties": {
        "payeeId": {
          "type": "number"
        }
      }
    },
    "disableWalletResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "format": "string"
        },
        "data": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "disableWalletErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": false
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
    "enableWallet": {
      "properties": {
        "payeeId": {
          "type": "number"
        }
      }
    },
    "enableWalletResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "format": "string"
        },
        "data": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "enableWalletErrorResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": false
        },
        "message": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
  }
}